Defined project build environment in a Dockerfile
The docker image is linked to the project directory through the
volume '/project'.

When running inside the container as the root user, the files and
directories generated by the build process are owned by root. But
this fact makes the files on the host side have the uid:gid pair
set to 0:0, which make them owned by root even inside the host.
The docker USER directive let the user running in the container to
be specified as an option of the run command. For example

    docker run -u `id -u`:`id -g` ...

runs the container with the same uid and gid of the host user.

However, the user does not exist inside the container, so the JVM
cannot recognize where the user home directory is. Both user name
and user home properties are resolved to '?', causing the creation
of a '?' folder inside the project directory during the build
process. To overcome this issue, the user.home property is
explicitly set in the _JAVA_OPTIONS environment variable.