Gradle integration testing for initSite task
Gradle test builds are run in a separate JVM, therefore an execution
classpath must be set externally. The classpath is passed from the
gradle build script by adding a system property named

    plugin.test.classpath

that contains the plugin's runtime classpath. The test classes can
access this property and pass it to the gradle runner.